fix/Расширил компонент. Вынес кнопку в отдельный компонент
