+++ src/components/ui/input-file/index.ts
@@ -1,2 +1,3 @@
 export * from './input-file';
 export * from './input-file-container/input-file-container';
+export * from './input-file-button/input-file-button';
