+++ src/components/ui/input-file/input-file-button/input-file-button.module.css
@@ -1,0 +1,3 @@
+.input {
+  display: none;
+}
