+++ src/components/ui/input-file/input-file.module.css
@@ -12,7 +12,3 @@
   margin-bottom: 14px;
   color: #eb5757;
 }
-
-.input {
-  display: none;
-}
