+++ src/components/ui/input-file/input-file.tsx
@@ -1,13 +1,11 @@
-import React, { useRef, useState, useEffect, FC, useCallback } from 'react';
+import React, { useRef, useState, useEffect, FC } from 'react';
 import cn from 'classnames';
-
-// import { Icon } from '../icon';
 
 import styles from './input-file.module.css';
 
 // Компоненты
-import { Button } from '../Button/index';
 import { InputFileContainer } from './index';
+import { InputFileButton } from './index';
 
 interface IInputFileProps {
   typesListFiles?: string[],
@@ -17,8 +15,6 @@ interface IInputFileProps {
 export const InputFile: FC<IInputFileProps> = ({ cb, typesListFiles }): JSX.Element => {
   const [ file, setFile ] = useState<null | File>(null);
 
-  // Текст кнопки
-  const [ textButton, setTextButton ] = useState('Добавить файл');
   // Имя файла для отображения
   const [ nameFile, setNameFile ] = useState('');
   // Текст для ошибки
@@ -49,39 +45,12 @@ export const InputFile: FC<IInputFileProps> = ({ cb, typesListFiles }): JSX.Elem
     }
   }, [ typesListFiles ]);
 
-  // Добавляю текст к кнопке
-  useEffect(() => {
-    setTextButton(nameFile ? 'Заменить файл' : 'Добавить файл');
-  }, [ nameFile ]);
-
   // Добавляю текст к сообщению ою ошибке
   useEffect(() => {
     nameFile ? 
       setMessageError(file ? '' : 'Файл содержит кириллицу, пожалуйста, переименуйте его.') : 
       setMessageError('');
   }, [ file, nameFile ]);
-
-  // Вызывает у input file его событие по умолчанию
-  const fileOnClick = useCallback((): void => {
-    if (inputRef.current) {
-      const input: HTMLInputElement = inputRef.current;
-      input.click();
-    }
-  }, [ inputRef ]);
-
-  // Когда файл выбран
-  const handlerChange = (e: React.ChangeEvent<HTMLInputElement>): void => {
-    const input: HTMLInputElement = e.target;
-    if (input.files) {
-      const file: File = input.files[0];
-      const isEnglishName = /^[\w]+.[\w]+$/;
-
-      const isValidName: boolean = isEnglishName.test(file.name);
-      isValidName ? setFile(file) : setFile(null);
-
-      setNameFile(file.name);
-    }
-  };
 
   return (
     <div className={ cn(styles.addFile) }>
@@ -100,23 +69,14 @@ export const InputFile: FC<IInputFileProps> = ({ cb, typesListFiles }): JSX.Elem
           </p>
         )
       }
-      <input
-        type='file'
-        ref={ inputRef }
-        onChange={ handlerChange }
-        className={ cn(styles.input) }
-      />
-      <Button 
-        label={ textButton } 
-        onClick={ fileOnClick } 
-        size={ 's' } 
-        icon={ 'plus' } 
-        iconPlace={ 'left' } 
-        border={ 'bottomLeft' } 
-        width={ '182px' }
-        align={ 'start'}
-        gap={'3px'}
-      />
+      {
+        <InputFileButton 
+          inputRef={ inputRef } 
+          nameFile={ nameFile } 
+          setFile={ setFile } 
+          setNameFile={ setNameFile } 
+        />
+      }
     </div>
   );
 };
